+++ app/scripts/lib/lforms-data.js
@@ -177,15 +177,80 @@ if (typeof LForms === 'undefined')
       this.PATH_DELIMITER = data.PATH_DELIMITER || "/";
       this.answerLists = data.answerLists;
       this.copyrightNotice = data.copyrightNotice;
-      this.fhirVersion = data.fhirVersion;
 
       // when the skip logic rule says the form is done
       this._formDone = false;
+
+      if (LForms.FHIR)
+        this._initializeFormFHIRData(data);
 
       // update internal data (_id, _idPath, _codePath, _displayLevel_),
       // that are used for widget control and/or for performance improvement.
       this._initializeInternalData();
 
+    },
+
+
+    /**
+     *  Registers a callback for use if the rendering of the form
+     *  requires the asynchronous operations (e.g. the loading of external
+     *  resources) which could affect the content.  Potentially, the function could
+     *  be called more than once, or after a related group of resources have
+     *  completed loading.
+     */
+    addAsyncChangeListener: function(listener) {
+      if (!this._asyncChangeListeners)
+        this._asyncChangeListeners = [];
+      this._asyncChangeListeners.push(listener);
+    },
+
+    /**
+     *  Calls the listeners registered via addAsyncChangeListener.
+     */
+    _notifyAsyncChangeListeners: function() {
+      if (this._asyncChangeListeners) {
+        for (var i=0, len=this._asyncChangeListeners.length; i<len; ++i)
+          this._asyncChangeListeners[i]();
+      }
+    },
+
+
+    /**
+     *  Initializes form-level FHIR data.
+     * @param an LForms form definition object (or LFormsData).
+     */
+    _initializeFormFHIRData: function(data) {
+      var lfData = this;
+      this._asyncLoadCounter = 0;
+      this.fhirVersion = data.fhirVersion;
+      this._fhirpathVars = {};
+      this.fhirExtensions = data.fhirExtensions;
+      if (LForms.fhirContext) {
+        var contextItems = LForms.Util.findObjectInArray(this.fhirExtensions, 'url',
+          "http://hl7.org/fhir/StructureDefinition/questionnaire-launchContext", 0, true);
+        for (var i=0, len=contextItems.length; i<len; ++i) {
+          var contextItemExt = contextItems[i].extension;
+          var name;
+          var typeList = [];
+          for (var j=0, jLen=contextItemExt.length; j<jLen; ++j) {
+            var fieldExt = contextItemExt[j];
+            if (!name && fieldExt.url === 'name')
+              name = fieldExt.valueId;
+            else if (fieldExt.url === 'type')
+              typeList.push(fieldExt.valueCode)
+          }
+          if (name && typeList.length > 0) {
+            this._asyncLoadCounter++;
+            LForms.fhirContext.getCurrent(typeList, function(resource) {
+              lfData._asyncLoadCounter--;
+              if (resource)
+                lfData._fhirpathVars[name] = resource;
+              if (lfData._asyncLoadCounter === 0)
+                lfData._notifyAsyncChangeListeners()
+            });
+          }
+        }
+      }
     },
 
 
@@ -324,10 +389,34 @@ if (typeof LForms === 'undefined')
               item._calculatedExprExt.valueExpression.language=="text/fhirpath") {
             if (!questResp)
               questResp = fhir.SDC.convertLFormsToQuestionnaireResponse(lfData);
-            item.value = fhir.fhirpath.evaluate(questResp,
-              item._calculatedExprExt.valueExpression.expression);
+            var fhirPathVal = fhir.fhirpath.evaluate(questResp,
+              item._calculatedExprExt.valueExpression.expression, lfData._fhirpathVars);
+            this._setItemValueFromFHIRPath(item, fhirPathVal);
           }
         }
+      }
+    },
+
+
+    /**
+     *  Assigns the given FHIRPath result to the given item.
+     * @param item the item from the LFormsData object that is receiving the new
+     *  value.
+     * @param fhirPathRes the result of a FHIRPath evaluation.
+     */
+    _setItemValueFromFHIRPath: function(item, fhirPathRes) {
+      var fhirPathVal = fhirPathRes[0];
+      if (!fhirPathVal)
+        item.value = undefined;
+      else {
+        if (item.dataType === this._CONSTANTS.DATA_TYPE.DT) {
+          var d = new Date(fhirPathVal);
+          // Convert to local time, so the date does not get shifted for negative
+          // local timezones.
+          item.value = new Date(d.getUTCFullYear(), d.getUTCMonth(), d.getUTCDate());
+        }
+        else
+          item.value = fhirPathVal; // TBD: handle other types - Coding, etc.
       }
     },