+++ test/karma/fhir-sdc.spec.js
@@ -100,7 +100,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
         describe('Questionnaire to lforms item conversion', function () {
           it('should convert FHTData to lforms', function () {
-            var fhirQ = LForms.Util.convertLFormsToFHIRData('Questionnaire', fhirVersion, angular.copy(FHTData));
+            var fhirQ = LForms.Util.getFormFHIRData('Questionnaire', fhirVersion, angular.copy(FHTData));
             var convertedLfData = LForms.Util.convertFHIRQuestionnaireToLForms(fhirQ, fhirVersion);
 
             assert.equal(convertedLfData.name, 'USSG-FHT, (with mock-up items for skip logic demo)');
@@ -165,7 +165,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
           });
 
           it('should convert restrictions', function () {
-            var fhirQ = LForms.Util.convertLFormsToFHIRData('Questionnaire', fhirVersion, angular.copy(validationTestForm));
+            var fhirQ = LForms.Util.getFormFHIRData('Questionnaire', fhirVersion, angular.copy(validationTestForm));
             var convertedLfData = LForms.Util.convertFHIRQuestionnaireToLForms(fhirQ, fhirVersion);
 
             assert.equal(convertedLfData.items.length, 33);
@@ -193,7 +193,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
           it('should convert to SDC Questionnaire with extensions', function() {
             var fhirQ = fhir.SDC.convertLFormsToQuestionnaire(new LForms.LFormsData(angular.copy(FHTData)));
 
-            assert.equal(fhirQ.meta.profile[0], fhir.SDC.sdcQProfile);
+            assert.equal(fhirQ.meta.profile[0], fhir.SDC.QProfile);
             assert.equal(fhirQ.item[0].item[1].extension[0].url, "http://hl7.org/fhir/StructureDefinition/questionnaire-minOccurs");
             assert.equal(fhirQ.item[0].item[1].extension[1].url, "http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl");
 
@@ -263,14 +263,14 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
         describe('LForms data to QuestionnaireResponse conversion', function() {
 
           it('should convert to SDC Questionnaire with extensions', function() {
-            var fhirQR = LForms.Util.convertLFormsToFHIRData('QuestionnaireResponse', fhirVersion, angular.copy(FHTData));
+            var fhirQR = LForms.Util.getFormFHIRData('QuestionnaireResponse', fhirVersion, angular.copy(FHTData));
 
-            assert.equal(fhirQR.meta.profile[0], "http://hl7.org/fhir/us/sdc/StructureDefinition/sdc-questionnaireresponse");
+            assert.equal(fhirQR.meta.profile[0], fhir.SDC.QRProfile);
 
           });
 
           it('should convert to standard QuestionnaireResponse without any extensions', function() {
-            var fhirQR = LForms.Util.convertLFormsToFHIRData(
+            var fhirQR = LForms.Util.getFormFHIRData(
               'QuestionnaireResponse', fhirVersion, angular.copy(FHTData),
               {noExtensions: true});